moved env generation to env_utils.py
