+++ .gitignore
@@ -1,7 +1,9 @@
 *.csv
 *.pyc
-hbaselines.egg*
+*h_baselines.egg*
 __pycache__
 *zip
 run_tests.sh
 debug
+*.idea/
+*.ipynb_checkpoints/
