+++ tests/fast_tests/test_utils.py
@@ -6,7 +6,7 @@ from gym.spaces import Box
 
 from hbaselines.utils.train import parse_options, get_hyperparameters
 from hbaselines.utils.reward_fns import negative_distance
-from hbaselines.utils.misc import get_manager_ac_space, get_state_indices
+from hbaselines.utils.env_util import get_manager_ac_space, get_state_indices
 from hbaselines.utils.tf_util import gaussian_likelihood
 from hbaselines.goal_conditioned.td3 import GoalConditionedPolicy
 from hbaselines.multi_fcnet.td3 import MultiFeedForwardPolicy
@@ -290,23 +290,6 @@ class TestMisc(unittest.TestCase):
             expected_size=1,
         )
 
-        # test for ring1
-        ac_space = get_manager_ac_space(env_name="ring1", **params)
-        test_space(
-            ac_space,
-            expected_min=np.array([0 for _ in range(1)]),
-            expected_max=np.array([1 for _ in range(1)]),
-            expected_size=1,
-        )
-
-        ac_space = get_manager_ac_space(env_name="ring1", **rel_params)
-        test_space(
-            ac_space,
-            expected_min=np.array([-0.5 for _ in range(1)]),
-            expected_max=np.array([0.5 for _ in range(1)]),
-            expected_size=1,
-        )
-
         # test for merge0
         ac_space = get_manager_ac_space(env_name="merge0", **params)
         test_space(
@@ -409,30 +392,10 @@ class TestMisc(unittest.TestCase):
             expected_size=14,
         )
 
-        # test for grid0
-        ac_space = get_manager_ac_space(env_name="grid0", **params)
-        del ac_space  # TODO
-
-        # test for grid1
-        ac_space = get_manager_ac_space(env_name="grid1", **params)
-        del ac_space  # TODO
-
-        # test for bottleneck0
-        ac_space = get_manager_ac_space(env_name="bottleneck0", **params)
-        del ac_space  # TODO
-
-        # test for bottleneck1
-        ac_space = get_manager_ac_space(env_name="bottleneck1", **params)
-        del ac_space  # TODO
-
-        # test for bottleneck2
-        ac_space = get_manager_ac_space(env_name="bottleneck2", **params)
-        del ac_space  # TODO
-
     def test_state_indices(self):
         # non-relevant parameters for most tests
         params = dict(
-            ob_space=Box(-1, 1, shape=(2,)),
+            ob_space=Box(-1, 1, shape=(2,), dtype=np.float32),
             use_fingerprints=False,
             fingerprint_dim=1,
         )
@@ -476,12 +439,6 @@ class TestMisc(unittest.TestCase):
             [0]
         )
 
-        # test for ring1
-        self.assertListEqual(
-            get_state_indices(env_name="ring1", **params),
-            [0]
-        )
-
         # test for merge0
         self.assertListEqual(
             get_state_indices(env_name="merge0", **params),
@@ -517,26 +474,6 @@ class TestMisc(unittest.TestCase):
             get_state_indices(env_name="figureeight2", **params),
             [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13]
         )
-
-        # test for grid0
-        state_indices = get_state_indices(env_name="grid0", **params)
-        del state_indices  # TODO
-
-        # test for grid1
-        state_indices = get_state_indices(env_name="grid1", **params)
-        del state_indices  # TODO
-
-        # test for bottleneck0
-        state_indices = get_state_indices(env_name="bottleneck0", **params)
-        del state_indices  # TODO
-
-        # test for bottleneck1
-        state_indices = get_state_indices(env_name="bottleneck1", **params)
-        del state_indices  # TODO
-
-        # test for bottleneck2
-        state_indices = get_state_indices(env_name="bottleneck2", **params)
-        del state_indices  # TODO
 
 
 class TestTFUtil(unittest.TestCase):
